pyproject.toml
[tool.poetry]
name = "cat_uncertainty"
version = "0.1.0"
description = "A Python package for performing uncertainty quantification on catalytic data."
authors = ["Osman Mamun <mamun.che06@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "cat_uncertainty"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
ase = "^3.23.0"
requests = "^2.32.3"
fairchem-core = "^1.4.0"
torch-geometric = "^1.7.0"
pytorch-lightning = "^2.5.0.post0"
wandb = "^0.19.1"
torchmetrics = "^1.6.0"
dscribe = "^2.1.1"
pip = "^24.3.1"
torch-sparse = {version = "^0.6.18+pt24cu124", source = "pyg"}
torch-scatter = {version = "^2.1.2+pt24cu124", source = "pyg"}
ray = {extras = ["tune"], version = "^2.40.0"}
optuna = "^4.1.0"
rich = "^13.9.4"
gpytorch = "^1.13"
ipython = "^8.31.0"
seaborn = "^0.13.2"
pyro-ppl = "^1.9.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
autoflake = "^2.3.1"
ruff = "^0.9.2"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[[tool.poetry.source]]
name = "pyg"
url = "https://data.pyg.org/whl/torch-2.4.0+cu124.html"
priority = "supplemental"

[build-system]
requires = ["poetry-core", "setuptools", "wheel", "ninja"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 80
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes (includes F401 for unused imports)
    "I",   # isort
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
